line notes

+++ SQL/paradise_schema.sql
@@ -251,7 +251,7 @@ CREATE TABLE `player` (
   `lastchangelog` varchar(32) NOT NULL,
   `space_parallax` smallint(4) DEFAULT '1',
   `space_dust` smallint(4) DEFAULT '1',
-  `parallax_speed` smallint(4) DEFAULT '2',
+  `parallax_speed` float(24) DEFAULT '2',
   PRIMARY KEY (`id`),
   UNIQUE KEY `ckey` (`ckey`)
 ) ENGINE=InnoDB AUTO_INCREMENT=32446 DEFAULT CHARSET=latin1;
